Function List Functionality
Added the ability to delete and redraw specific functions. The function list now tracks the colors of the functions as they are (re)drawn and I seem to have worked out most of the glitches involving the functions not being properly deleted or a previous version of the function still existing after deletion.